was queued

+++ dbos/core.py
@@ -156,6 +156,7 @@ def _init_workflow(
             json.dumps(ctx.authenticated_roles) if ctx.authenticated_roles else None
         ),
         "assumed_role": ctx.assumed_role,
+        "was_queued": queue is not None,
     }
 
     # If we have a class name, the first arg is the instance and do not serialize
@@ -188,7 +189,8 @@ def _execute_workflow(
         output = func(*args, **kwargs)
         status["status"] = "SUCCESS"
         status["output"] = utils.serialize(output)
-        dbos._sys_db.remove_from_queue(status["workflow_uuid"])
+        if status["was_queued"]:
+            dbos._sys_db.remove_from_queue(status["workflow_uuid"])
         dbos._sys_db.buffer_workflow_status(status)
     except DBOSWorkflowConflictIDError:
         # Retrieve the workflow handle and wait for the result.
@@ -201,7 +203,8 @@ def _execute_workflow(
     except Exception as error:
         status["status"] = "ERROR"
         status["error"] = utils.serialize(error)
-        dbos._sys_db.remove_from_queue(status["workflow_uuid"])
+        if status["was_queued"]:
+            dbos._sys_db.remove_from_queue(status["workflow_uuid"])
         dbos._sys_db.update_workflow_status(status)
         raise
